refactor(Movement): dao -> domain.position 패키지 이동
